Example of an experiment config file

+++ examples/generate-pattern.cpp
@@ -6,8 +6,7 @@ using namespace std;
 
 int main(int argc, char** argv)
 {
-    if (argc <= 3) 
-    {
+    if (argc <= 3) {
         cout << "Usage: " << argv[0] << " file count length [charset]" << endl;
         cout << " - file:       The data to extract the patterns from" << endl;
         cout << " - count:      Number of patterns to extract" << endl;
@@ -26,7 +25,7 @@ int main(int argc, char** argv)
     size_t count = atoi(argv[2]);
     size_t length = atoi(argv[3]);
     int charset = argc <= 4 ? 0 : atoi(argv[4]);
-    
+
     // CHARSET
     auto is_special_char = [](char c) { return isspace(c) || ispunct(c); };
 
@@ -36,86 +35,77 @@ int main(int argc, char** argv)
     };
     auto charset_predicate = charset_predicates[charset];
     auto charset_string_predicate = [&](string s) { return all_of(s.begin(), s.end(), charset_predicate); };
-    
+
     // load (cached) CST
     string cst_file = file + ".sdsl";
     cst_sct3<> cst;
-    if (!load_from_file(cst, cst_file))
-    {
+    if (!load_from_file(cst, cst_file)) {
         construct(cst, file, 1);
         store_to_file(cst, cst_file);
     }
 
     // traverse
     size_t size = cst.size();
+    // TODO: replace by priority_queue<  > min queue
     stack<pair<size_t, string>> found;
     auto min_occ = [&]() { return found.empty() ? 0 : found.top().first; };
-    for (auto it=cst.begin(); it!=cst.end(); ++it) 
-    {
-        if (it.visit() == 1) 
-        {
+    for (auto it=cst.begin(); it!=cst.end(); ++it) {
+        if (it.visit() == 1) {
             auto v = *it;
             auto depth = cst.depth(v);
             if (depth == 0)
                 continue;
-            
+
             auto curr_min_occ = min_occ();
             auto curr_occ = cst.size(v);
-            if (curr_occ <= curr_min_occ && found.size() == count)
-            {
+            if (curr_occ <= curr_min_occ && found.size() == count) {
                 it.skip_subtree();
                 continue;
             }
-            
+
             auto begin = cst.csa[cst.lb(v)];
-            string phrase = extract(cst.csa, begin, begin + min(depth, length) - 1);
-                        
+            string phrase = extract(cst.csa, begin, begin + min((size_t)depth, (size_t)length) - 1);
+
             // check charset
-            if (!charset_string_predicate(phrase))
-            {
+            if (!charset_string_predicate(phrase)) {
                 it.skip_subtree();
                 continue;
             }
-            
+
             // check length
-            if (depth >= length)
-            {
+            if (depth >= length) {
                 // check for word boundary (approximation)
-                if (!is_special_char(cst.csa.L[cst.rb(v)]) 
-                 && !is_special_char(cst.csa.L[cst.lb(v)]))
-                {
+                if (!is_special_char(cst.csa.L[cst.rb(v)])
+                    && !is_special_char(cst.csa.L[cst.lb(v)])) {
                     it.skip_subtree();
                     continue;
                 }
 
                 stack<pair<size_t, string>> helper;
-                while (!found.empty() && found.top().first < curr_occ)
-                {
+                while (!found.empty() && found.top().first < curr_occ) {
                     helper.push(found.top());
                     found.pop();
                 }
                 found.emplace(curr_occ, phrase);
-                while (found.size() < count && !helper.empty())
-                {
+                while (found.size() < count && !helper.empty()) {
                     found.push(helper.top());
                     helper.pop();
                 }
-                
+
                 // report status
                 cerr << "Found " << found.size() << " phrases with min. #occurrences of " << found.top().first << " so far" << endl;
-                
+
                 it.skip_subtree();
             }
         }
     }
-    
+
     // fill
     while (found.size() < count)
-        found.emplace(0, string(length, 'x'));    
+        found.emplace(0, string(length, 'x'));
 
     // output
-    while (!found.empty())
-    {
+    while (!found.empty()) {
         cout << found.top().second << endl;
         found.pop();
     }
